refactor advisory policy support actions
