+++ src/app/router_registry.py
@@ -17,6 +17,9 @@ from app.routers.advisory_policy_evaluation_actions import (
 from app.routers.advisory_policy_evaluation_evidence import (
     router as advisory_policy_evaluation_evidence_router,
 )
+from app.routers.advisory_policy_evaluation_support_actions import (
+    router as advisory_policy_evaluation_support_actions_router,
+)
 from app.routers.advisory_policy_evaluations import (
     router as advisory_policy_evaluations_router,
 )
@@ -228,6 +231,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(advisory_policy_evaluations_router)
     app.include_router(advisory_policy_evaluation_evidence_router)
     app.include_router(advisory_policy_evaluation_actions_router)
+    app.include_router(advisory_policy_evaluation_support_actions_router)
     app.include_router(proposal_generation_router)
     app.include_router(proposals_router)
     app.include_router(proposal_operations_router)
