+++ src/app/routers/advisory_policy_evaluation_support_actions.py
@@ -1,0 +1,65 @@
+from fastapi import APIRouter, Header
+
+from app.contracts.advisory_policy import (
+    AdvisoryPolicyBodyRequest,
+    AdvisoryPolicyEnvelopeResponse,
+)
+from app.middleware.correlation import correlation_id_var
+from app.routers.advisory_policy_evaluation_common import POLICY_EVALUATION_PATH
+from app.services.advisory_service_provider import advisory_policy_service
+
+router = APIRouter(prefix="/api/v1", tags=["advisory-policy"])
+
+
+@router.post(
+    "/advisory-policy-evaluations/{evaluation_id}/report-packages",
+    response_model=AdvisoryPolicyEnvelopeResponse,
+    summary="Request Advisory Policy Report Package",
+    description=(
+        "Requests a source-owned advisor/compliance policy sign-off package through lotus-advise. "
+        "Gateway does not promote blocked or degraded evaluations to client-ready publication."
+    ),
+)
+async def request_policy_report_package(
+    request: AdvisoryPolicyBodyRequest,
+    evaluation_id: str = POLICY_EVALUATION_PATH,
+    idempotency_key: str | None = Header(
+        default=None,
+        alias="Idempotency-Key",
+        description="Optional idempotency key for policy report packages.",
+        examples=["idem-policy-report-1"],
+    ),
+) -> AdvisoryPolicyEnvelopeResponse:
+    return await advisory_policy_service().request_policy_report_package(
+        evaluation_id=evaluation_id,
+        body=request.body,
+        idempotency_key=idempotency_key,
+        correlation_id=correlation_id_var.get(),
+    )
+
+
+@router.post(
+    "/advisory-policy-evaluations/{evaluation_id}/ai-evidence",
+    response_model=AdvisoryPolicyEnvelopeResponse,
+    summary="Request Advisory Policy AI Evidence",
+    description=(
+        "Requests bounded policy evidence through lotus-advise. AI output remains "
+        "non-authoritative; Advise owns redaction, fail-closed posture, and client-ready blockers."
+    ),
+)
+async def request_policy_ai_evidence(
+    request: AdvisoryPolicyBodyRequest,
+    evaluation_id: str = POLICY_EVALUATION_PATH,
+    idempotency_key: str | None = Header(
+        default=None,
+        alias="Idempotency-Key",
+        description="Optional idempotency key for policy AI evidence requests.",
+        examples=["idem-policy-ai-evidence-1"],
+    ),
+) -> AdvisoryPolicyEnvelopeResponse:
+    return await advisory_policy_service().request_policy_ai_evidence(
+        evaluation_id=evaluation_id,
+        body=request.body,
+        idempotency_key=idempotency_key,
+        correlation_id=correlation_id_var.get(),
+    )
